Move basic utility functions from ethUtils to main utils file
