--- src/ethUtils.h
@@ -1,74 +1,0 @@
-/*******************************************************************************
- *   Ledger Ethereum App
- *   (c) 2016-2019 Ledger
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- ********************************************************************************/
-
-#ifndef _ETHUTILS_H_
-#define _ETHUTILS_H_
-
-#include <stdint.h>
-
-#include "cx.h"
-#include "chainConfig.h"
-
-#define KECCAK256_HASH_BYTESIZE 32
-
-bool getEthAddressFromKey(cx_ecfp_public_key_t *publicKey, uint8_t *out, cx_sha3_t *sha3Context);
-
-bool getEthAddressStringFromKey(cx_ecfp_public_key_t *publicKey,
-                                char *out,
-                                cx_sha3_t *sha3Context,
-                                uint64_t chainId);
-
-bool u64_to_string(uint64_t src, char *dst, uint8_t dst_size);
-
-bool getEthAddressStringFromBinary(uint8_t *address,
-                                   char *out,
-                                   cx_sha3_t *sha3Context,
-                                   uint64_t chainId);
-
-bool getEthDisplayableAddress(uint8_t *in,
-                              char *out,
-                              size_t out_len,
-                              cx_sha3_t *sha3,
-                              uint64_t chainId);
-
-bool adjustDecimals(const char *src,
-                    size_t srcLength,
-                    char *target,
-                    size_t targetLength,
-                    uint8_t decimals);
-
-static __attribute__((no_instrument_function)) inline int allzeroes(const void *buf, size_t n) {
-    uint8_t *p = (uint8_t *) buf;
-    for (size_t i = 0; i < n; ++i) {
-        if (p[i]) {
-            return 0;
-        }
-    }
-    return 1;
-}
-static __attribute__((no_instrument_function)) inline int ismaxint(uint8_t *buf, int n) {
-    for (int i = 0; i < n; ++i) {
-        if (buf[i] != 0xff) {
-            return 0;
-        }
-    }
-    return 1;
-}
-
-static const char HEXDIGITS[] = "0123456789abcdef";
-
-#endif  // _ETHUTILS_H_
