+++ src/eth_plugin_handler.c
@@ -3,7 +3,6 @@
 #include "eth_plugin_internal.h"
 #include "shared_context.h"
 #include "network.h"
-#include "ethUtils.h"
 
 void eth_plugin_prepare_init(ethPluginInitContract_t *init,
                              const uint8_t *selector,
