+++ src/eth_plugin_internal.c
@@ -1,6 +1,5 @@
 #include <string.h>
 #include "eth_plugin_internal.h"
-#include "ethUtils.h"  // allzeroes
 
 bool erc20_plugin_available_check(void);
