+++ src_features/getPublicKey/cmd_getPublicKey.c
@@ -2,7 +2,6 @@
 #include "apdu_constants.h"
 #include "utils.h"
 #include "feature_getPublicKey.h"
-#include "ethUtils.h"
 #include "common_ui.h"
 #include "os_io_seproxyhal.h"
