+++ src_features/signMessageEIP712/field_hash.c
@@ -8,7 +8,6 @@
 #include "mem_utils.h"
 #include "shared_context.h"
 #include "ui_logic.h"
-#include "ethUtils.h"        // KECCAK256_HASH_BYTESIZE
 #include "context_712.h"     // contract_addr
 #include "utils.h"           // u64_from_BE
 #include "apdu_constants.h"  // APDU response codes
