+++ src_nbgl/ui_approve_tx.c
@@ -3,7 +3,6 @@
 #include "shared_context.h"
 #include "ui_callbacks.h"
 #include "ui_nbgl.h"
-#include "ethUtils.h"
 #include "ui_signing.h"
 #include "plugins.h"
 #include "domain_name.h"
