+++ src_plugins/erc1155/erc1155_ui.c
@@ -3,7 +3,6 @@
 #include <string.h>
 #include "erc1155_plugin.h"
 #include "eth_plugin_interface.h"
-#include "ethUtils.h"
 #include "utils.h"
 
 static void set_approval_for_all_ui(ethQueryContractUI_t *msg, erc1155_context_t *context) {
