+++ src_plugins/erc721/erc721_ui.c
@@ -3,7 +3,6 @@
 #include <string.h>
 #include "erc721_plugin.h"
 #include "eth_plugin_interface.h"
-#include "ethUtils.h"
 #include "utils.h"
 
 static void set_approval_ui(ethQueryContractUI_t *msg, erc721_context_t *context) {
